PIR management and reporting through the occupancy zigbee cluster
On every wakeup check the status of the PIR sensor. If the PIR
device is high report to the network that occupancy is true and
sleep (no deep sleep) until the PIR sensor goes low. At this
point if the user has set a 'PIR occupied to unoccupied delay',
setup a timer for N seconds, otherwise immediately report the
state as unoccupied.