Add an idle task and power management
The idle task does nothing for now other than powering down the device.

The power management policy is to power down as soon as the device is
idle and place it deep sleep mode. Deep sleep mode always results in
RAM memory being cleared when waking up, but it also reduces the total
power consumption of the device by around 2uA.

Detecting if the device is idle is a bit tricky. For now check if the
next timer event is the zigbee poll rate task, which signifies that
there are no other short term tasks to execute and enter deep sleep
for 6 hours.